// SDKs/googleapis.com/dialogflow/v3beta1/typescript/src/sdk/models/shared/googleclouddialogflowcxv3beta1responsemessagemixedaudio.ts
import { SpeakeasyMetadata, SpeakeasyBase } from "../../../internal/utils";
import { GoogleCloudDialogflowCxV3beta1ResponseMessageMixedAudioSegmentInput } from "./googleclouddialogflowcxv3beta1responsemessagemixedaudiosegment";
import { GoogleCloudDialogflowCxV3beta1ResponseMessageMixedAudioSegment } from "./googleclouddialogflowcxv3beta1responsemessagemixedaudiosegment";



// GoogleCloudDialogflowCxV3beta1ResponseMessageMixedAudioInput
/** 
 * Represents an audio message that is composed of both segments synthesized from the Dialogflow agent prompts and ones hosted externally at the specified URIs. The external URIs are specified via play_audio. This message is generated by Dialogflow only and not supposed to be defined by the user.
**/
export class GoogleCloudDialogflowCxV3beta1ResponseMessageMixedAudioInput extends SpeakeasyBase {
  @SpeakeasyMetadata({ data: "json, name=segments", elemType: GoogleCloudDialogflowCxV3beta1ResponseMessageMixedAudioSegmentInput })
  segments?: GoogleCloudDialogflowCxV3beta1ResponseMessageMixedAudioSegmentInput[];
}


// GoogleCloudDialogflowCxV3beta1ResponseMessageMixedAudio
/** 
 * Represents an audio message that is composed of both segments synthesized from the Dialogflow agent prompts and ones hosted externally at the specified URIs. The external URIs are specified via play_audio. This message is generated by Dialogflow only and not supposed to be defined by the user.
**/
export class GoogleCloudDialogflowCxV3beta1ResponseMessageMixedAudio extends SpeakeasyBase {
  @SpeakeasyMetadata({ data: "json, name=segments", elemType: GoogleCloudDialogflowCxV3beta1ResponseMessageMixedAudioSegment })
  segments?: GoogleCloudDialogflowCxV3beta1ResponseMessageMixedAudioSegment[];
}